loader_app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
chromadb==0.3.26
click==8.1.7
clickhouse-connect==0.8.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
duckdb==1.1.3
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
hnswlib==0.8.0
httptools==0.6.4
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
langchain==0.0.213
langchainplus-sdk==0.0.20
lz4==4.3.3
marshmallow==3.23.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.1
numpy==1.26.4
onnxruntime==1.19.2
openai==0.27.0
openapi-schema-pydantic==1.2.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
posthog==3.7.0
propcache==0.2.0
protobuf==5.28.3
pulsar-client==3.5.0
pydantic==1.10.19
pydantic_core==2.23.4
pypdf==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.4.0
tokenizers==0.20.3
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==0.24.0
websockets==14.1
yarl==1.17.1
zstandard==0.23.0
